upd 06.03.2023
- теперь рабоатет в 1080 разрешении
- проклик на отказ от резервной почты мейл
- с 5 на 10 сек время на создание мафайла у сда

+++ AutoSDA/Deserialize Classes/TzIDOnlineSim.cs
@@ -9,7 +9,7 @@ namespace AutoSDA.Deserialize_Classes
     internal class TzIDOnlineSim
     {
         public int country { get; set; }
-        public int sum { get; set; }
+        public double sum { get; set; }
         public string service { get; set; }
         public string number { get; set; }
         public string response { get; set; }
